pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openprinttag-pn5180-rpi"
version = "0.1.0"
description = "Small app that reads/write openprinttags using pn5180 on rpi."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cbor2>=5.4.0",
    "pyyaml>=6.0",
    "pn5180pi",
    "pydantic2>=1.1.15",
    "rpi-ws281x>=5.0.0",
    "adafruit-circuitpython-neopixel>=6.3.21",
    "paho-mqtt>=2.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv.sources]
pypn5180 = { git = "https://github.com/captainbeeheart/pypn5180" }

[tool.uv]
dev-dependencies = ["datamodel-code-generator"]
